Add KOMA script title commands
- add \extratitle, \subtitle and \subject
- move KOMA commands availabe in all KOMA classes into a separate 
package
- add tests to check, whether the package is loaded
